+++ pkgs/yaml/test/yaml_test.dart
@@ -8,7 +8,6 @@
 // ignore_for_file: avoid_dynamic_calls
 
 import 'package:test/test.dart';
-import 'package:yaml/src/error_listener.dart';
 import 'package:yaml/yaml.dart';
 
 import 'utils.dart';
@@ -392,7 +391,7 @@ void main() {
     test('[Example 2.13]', () {
       expectYamlLoads(cleanUpLiteral('''
         \\//||\\/||
-        // ||  ||__'''), '''
+        // ||  ||__\n'''), '''
         # ASCII Art
         --- |
           \\//||\\/||
@@ -400,7 +399,8 @@ void main() {
     });
 
     test('[Example 2.14]', () {
-      expectYamlLoads("Mark McGwire's year was crippled by a knee injury.", '''
+      expectYamlLoads(
+          "Mark McGwire's year was crippled by a knee injury.\n", '''
         --- >
           Mark McGwire's
           year was crippled
@@ -414,7 +414,7 @@ void main() {
           63 Home Runs
           0.288 Batting Average
 
-        What a year!'''), '''
+        What a year!\n'''), '''
         >
          Sammy Sosa completed another
          fine season with great stats.
@@ -429,7 +429,7 @@ void main() {
       expectYamlLoads({
         'name': 'Mark McGwire',
         'accomplishment': 'Mark set a major league home run record in 1998.\n',
-        'stats': '65 Home Runs\n0.278 Batting Average'
+        'stats': '65 Home Runs\n0.278 Batting Average\n'
       }, '''
         name: Mark McGwire
         accomplishment: >
@@ -715,17 +715,17 @@ void main() {
       test(
           'from \\r to \\n',
           () => expectYamlLoads(
-              ['foo\nbar'], indentLiteral('- |\n  foo\r  bar')));
+              ['foo\nbar\n'], indentLiteral('- |\n  foo\r  bar')));
       test(
           'from \\r\\n to \\n',
           () => expectYamlLoads(
-              ['foo\nbar'], indentLiteral('- |\n  foo\r\n  bar')));
+              ['foo\nbar\n'], indentLiteral('- |\n  foo\r\n  bar')));
     });
 
     test('[Example 5.11]', () {
       expectYamlLoads(cleanUpLiteral('''
         Line break (no glyph)
-        Line break (glyphed)'''), '''
+        Line break (glyphed)\n'''), '''
         |
           Line break (no glyph)
           Line break (glyphed)''');
@@ -1704,7 +1704,7 @@ void main() {
     test('[Example 8.21]', () {
       // The spec doesn't include a newline after "value" in the parsed map, but
       // the block scalar is clipped so it should be retained.
-      expectYamlLoads({'literal': 'value\n', 'folded': 'value'}, '''
+      expectYamlLoads({'literal': 'value\n', 'folded': 'value\n'}, '''
         literal: |2
           value
         folded: