+++ pkgs/yaml_edit/test/append_test.dart
@@ -265,5 +265,98 @@ a:
       expect(doc.toString(), equals('[0]'));
       expectYamlBuilderValue(doc, [0]);
     });
+
+    test('spanning multiple lines with trailing comma', () {
+      final doc = YamlEditor('''
+[
+  0,
+  1,
+]
+''');
+      doc.appendToList([], 2);
+      expect(doc.toString(), equals('''
+[
+  0,
+  1,
+  2,
+]
+'''));
+      expectYamlBuilderValue(doc, [0, 1, 2]);
+    });
+
+    test('spanning multiple lines with trailing comma and comment', () {
+      final doc = YamlEditor('''
+[
+  0,
+  1, # comment
+]
+''');
+      doc.appendToList([], 2);
+      expect(doc.toString(), equals('''
+[
+  0,
+  1, # comment
+  2,
+]
+'''));
+      expectYamlBuilderValue(doc, [0, 1, 2]);
+    });
+
+    test('nested flow list spanning multiple lines with trailing comma', () {
+      final doc = YamlEditor('''
+analyzer:
+  exclude:
+    [
+      foo/**,
+      build/**,
+    ]
+''');
+      doc.appendToList(['analyzer', 'exclude'], 'android/**');
+      expect(doc.toString(), equals('''
+analyzer:
+  exclude:
+    [
+      foo/**,
+      build/**,
+      android/**,
+    ]
+'''));
+      expectYamlBuilderValue(doc, {
+        'analyzer': {
+          'exclude': ['foo/**', 'build/**', 'android/**']
+        }
+      });
+    });
+
+    test('spanning multiple lines without trailing comma', () {
+      final doc = YamlEditor('''
+[
+  0,
+  1
+]
+''');
+      doc.appendToList([], 2);
+      expect(doc.toString(), equals('''
+[
+  0,
+  1
+, 2]
+'''));
+      expectYamlBuilderValue(doc, [0, 1, 2]);
+    });
+
+    test('single line with trailing comma', () {
+      final doc = YamlEditor('[0, 1, ]');
+      doc.appendToList([], 2);
+      expect(doc.toString(), equals('[0, 1, 2,]'));
+      expectYamlBuilderValue(doc, [0, 1, 2]);
+    });
+
+    test('single line with trailing comma (no space)', () {
+      final doc = YamlEditor('[0,1,]');
+      doc.appendToList([], 2);
+      expect(doc.toString(), equals('[0,1, 2,]'));
+      expectYamlBuilderValue(doc, [0, 1, 2]);
+    });
   });
 }